simplify very very short functions
These functions are implemented to use Gnu Scientific Library (GSL)
which is written in C. GSL has a root solver which finds a root of
user-defined 1D functions and greens_functions depends on it. Since
C does not have generics or function objects, the GSL root solver
recieves function pointer which has interface like `f(double, void*)`.
With this, you can pass all the parameters by packing them into one
struct and cast a pointer to the struct to `void*`. This is the reason
why greens_functions often defines auxiliary functions that casts
`void*` to the struct and pass the data contained it to the function
that actually does calculation. In this commit, these functions are
simplified to make them shorter.